Made function move_pdf in loader. See desc for more detail
move_pdf takes a filepath as an input, and takes any PDF files inside of it, and puts them into the data/raw_PDFs directory. It also returns the name of the directory they have been copied into, the names of the PDF files, and how many were copied - as a check.